+++ src/main/kotlin/blackjack/state/BlackjackState.kt
@@ -1,11 +1,11 @@
 package blackjack.state
 
-sealed interface State {
-    sealed interface Running : State {
+sealed interface BlackjackState {
+    sealed interface Running : BlackjackState {
         data object Hit : Running
     }
 
-    sealed interface Finished : State {
+    sealed interface Finished : BlackjackState {
         data object Bust : Finished
 
         data object Stay : Finished
